+++ examples/langchain_async_demo.py
@@ -7,7 +7,7 @@ from langchain.callbacks.streaming_stdout import StreamingStdOutCallbackHandler
 from mli import LangchainMLIClient
 
 
-ENDPOINT = 'http://127.0.0.1:5000'
+ENDPOINT = 'http://127.0.0.1:5000/api/1.0'
 
 
 async def langchain_async_demo():