+++ mli/server.py
@@ -559,10 +559,10 @@ class MLIServer:
                     elif msg.type == WSMsgType.ERROR:
                         print(f'[ERROR] websocket closed with exception: {ws.exception()}')
         except ExceptionGroup as e:
-            traceback.print_exc()
+            # traceback.print_exc()
             print(f'[ERROR] websocket ExceptionGroup: {e}')
         except Exception as e:
-            traceback.print_exc()
+            # traceback.print_exc()
             print(f'[ERROR] TaskGroup Exception: {e}')
 
         if ws in self.ws_proc_map: